+++ src/schedule/resolve_state.rs
@@ -1,4 +1,4 @@
-//! Configure [`ResolveStateSet`] system sets.
+//! Configure [`ResolveStateSystems`] system sets.
 
 #[cfg(feature = "bevy_app")]
 pub use app::*;
@@ -13,13 +13,13 @@ mod app {
 
     use crate::state::State;
 
-    use super::{ResolveStateSet, schedule_resolve_state};
+    use super::{ResolveStateSystems, schedule_resolve_state};
 
-    /// A plugin that configures the [`ResolveStateSet<S>`] system sets for the [`State`]
+    /// A plugin that configures the [`ResolveStateSystems<S>`] system sets for the [`State`]
     /// type `S` in the [`StateFlush`](crate::schedule::StateFlush) schedule.
     ///
     /// To specify a dependency relative to another `State` type `T`, add
-    /// [`ResolveStateSet::<T>::Resolve`] to [`after`](Self::after) or [`before`](Self::before).
+    /// [`ResolveStateSystems::<T>::Resolve`] to [`after`](Self::after) or [`before`](Self::before).
     ///
     /// Calls [`schedule_resolve_state<S>`].
     pub struct ResolveStatePlugin<S: State> {
@@ -60,13 +60,13 @@ mod app {
 
         /// Configure a `.after` system set.
         pub fn after<T: State>(mut self) -> Self {
-            self.after.push(ResolveStateSet::<T>::Resolve.intern());
+            self.after.push(ResolveStateSystems::<T>::Resolve.intern());
             self
         }
 
         /// Configure a `.before` system set.
         pub fn before<T: State>(mut self) -> Self {
-            self.before.push(ResolveStateSet::<T>::Resolve.intern());
+            self.before.push(ResolveStateSystems::<T>::Resolve.intern());
             self
         }
     }
@@ -78,7 +78,7 @@ use bevy_ecs::schedule::{
     Condition, InternedSystemSet, IntoScheduleConfigs as _, Schedule, SystemSet,
 };
 
-use crate::{schedule::ApplyFlushSet, state::State};
+use crate::{schedule::ApplyFlushSystems, state::State};
 
 /// A suite of system sets in the [`StateFlush`](crate::schedule::StateFlush)
 /// schedule for each [`State`] type `S`.
@@ -87,7 +87,7 @@ use crate::{schedule::ApplyFlushSet, state::State};
 /// [`ResolveStatePlugin<S>`] as follows:
 ///
 /// 1. [`Resolve`](Self::Resolve) (before or after other `Resolve` system sets based on
-///    state dependencies, and before [`ApplyFlushSet`])
+///    state dependencies, and before [`ApplyFlushSystems`])
 ///     1. [`Compute`](Self::Compute)
 ///     2. [`Trigger`](Self::Trigger)
 ///     3. [`Flush`](Self::Flush) (and [`AnyFlush`](Self::AnyFlush) if the global state will flush)
@@ -97,7 +97,7 @@ use crate::{schedule::ApplyFlushSet, state::State};
 ///         3. [`Enter`](Self::Enter) (and [`AnyEnter`](Self::AnyEnter) if the global state will
 ///            enter)
 #[derive(SystemSet)]
-pub enum ResolveStateSet<S: State> {
+pub enum ResolveStateSystems<S: State> {
     /// Resolve the state flush logic for `S`.
     Resolve,
     /// Optionally compute the next value for `S`.
@@ -124,7 +124,7 @@ pub enum ResolveStateSet<S: State> {
     _PhantomData(PhantomData<S>, Infallible),
 }
 
-impl<S: State> Clone for ResolveStateSet<S> {
+impl<S: State> Clone for ResolveStateSystems<S> {
     fn clone(&self) -> Self {
         match self {
             Self::Resolve => Self::Resolve,
@@ -143,21 +143,21 @@ impl<S: State> Clone for ResolveStateSet<S> {
     }
 }
 
-impl<S: State> PartialEq for ResolveStateSet<S> {
+impl<S: State> PartialEq for ResolveStateSystems<S> {
     fn eq(&self, other: &Self) -> bool {
         core::mem::discriminant(self) == core::mem::discriminant(other)
     }
 }
 
-impl<S: State> Eq for ResolveStateSet<S> {}
+impl<S: State> Eq for ResolveStateSystems<S> {}
 
-impl<S: State> Hash for ResolveStateSet<S> {
+impl<S: State> Hash for ResolveStateSystems<S> {
     fn hash<H: core::hash::Hasher>(&self, state: &mut H) {
         core::mem::discriminant(self).hash(state);
     }
 }
 
-impl<S: State> Debug for ResolveStateSet<S> {
+impl<S: State> Debug for ResolveStateSystems<S> {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         match self {
             Self::Resolve => write!(f, "Resolve"),
@@ -176,10 +176,10 @@ impl<S: State> Debug for ResolveStateSet<S> {
     }
 }
 
-/// Configure [`ResolveStateSet<S>`] for the [`State`] type `S` in a schedule.
+/// Configure [`ResolveStateSystems<S>`] for the [`State`] type `S` in a schedule.
 ///
 /// To specify a dependency relative to another `State` type `T`, include
-/// [`ResolveStateSet::<T>::Resolve`] in `after` or `before`.
+/// [`ResolveStateSystems::<T>::Resolve`] in `after` or `before`.
 ///
 /// Used in [`ResolveStatePlugin<S>`].
 pub fn schedule_resolve_state<S: State>(
@@ -189,45 +189,45 @@ pub fn schedule_resolve_state<S: State>(
 ) {
     // External ordering
     for &system_set in after {
-        schedule.configure_sets(ResolveStateSet::<S>::Resolve.after(system_set));
+        schedule.configure_sets(ResolveStateSystems::<S>::Resolve.after(system_set));
     }
     for &system_set in before {
-        schedule.configure_sets(ResolveStateSet::<S>::Resolve.before(system_set));
+        schedule.configure_sets(ResolveStateSystems::<S>::Resolve.before(system_set));
     }
 
     // Internal ordering
     schedule.configure_sets((
-        ResolveStateSet::<S>::Resolve.before(ApplyFlushSet),
+        ResolveStateSystems::<S>::Resolve.before(ApplyFlushSystems),
         (
-            ResolveStateSet::<S>::Compute,
+            ResolveStateSystems::<S>::Compute,
             // Logic in this system set should only run if not triggered.
-            ResolveStateSet::<S>::Trigger,
+            ResolveStateSystems::<S>::Trigger,
             // Logic in this system set should only run if triggered.
-            ResolveStateSet::<S>::Flush,
+            ResolveStateSystems::<S>::Flush,
         )
             .chain()
-            .in_set(ResolveStateSet::<S>::Resolve),
+            .in_set(ResolveStateSystems::<S>::Resolve),
         (
-            ResolveStateSet::<S>::Exit,
-            ResolveStateSet::<S>::Trans,
-            ResolveStateSet::<S>::Enter,
+            ResolveStateSystems::<S>::Exit,
+            ResolveStateSystems::<S>::Trans,
+            ResolveStateSystems::<S>::Enter,
         )
             .chain()
-            .in_set(ResolveStateSet::<S>::Flush),
-        ResolveStateSet::<S>::AnyFlush
+            .in_set(ResolveStateSystems::<S>::Flush),
+        ResolveStateSystems::<S>::AnyFlush
             .run_if(S::is_triggered)
-            .in_set(ResolveStateSet::<S>::Flush),
+            .in_set(ResolveStateSystems::<S>::Flush),
         (
-            ResolveStateSet::<S>::AnyExit
+            ResolveStateSystems::<S>::AnyExit
                 .run_if(S::is_enabled)
-                .in_set(ResolveStateSet::<S>::Exit),
-            ResolveStateSet::<S>::AnyTrans
+                .in_set(ResolveStateSystems::<S>::Exit),
+            ResolveStateSystems::<S>::AnyTrans
                 .run_if(S::is_enabled.and(S::will_be_enabled))
-                .in_set(ResolveStateSet::<S>::Trans),
-            ResolveStateSet::<S>::AnyEnter
+                .in_set(ResolveStateSystems::<S>::Trans),
+            ResolveStateSystems::<S>::AnyEnter
                 .run_if(S::will_be_enabled)
-                .in_set(ResolveStateSet::<S>::Enter),
+                .in_set(ResolveStateSystems::<S>::Enter),
         )
-            .in_set(ResolveStateSet::<S>::AnyFlush),
+            .in_set(ResolveStateSystems::<S>::AnyFlush),
     ));
 }