+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - **Added `no_std` support**
 - Added `NextStateStackCommandsExt` extension trait
 - Moved `StateFlush` to before `PreUpdate` schedule
+- Renamed `ApplyFlushSet` -> `ApplyFlushSystems` system set
+- Renamed `ResolveStateSet` -> `ResolveStateSystems` system set
 
 # Version 0.3.0